TradingTips/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
yfinance>=0.2.18
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit-aggrid>=0.3.4
ta>=0.10.2
numpy>=1.24.0
